+++ shuup/importer/importing/__init__.py
@@ -6,12 +6,13 @@
 # This source code is licensed under the OSL-3.0 license found in the
 # LICENSE file in the root directory of this source tree.
 
-from .importing import DataImporter
+from .importing import DataImporter, ImporterExampleFile
 from .meta import ImportMetaBase
 from .session import DataImporterRowSession
 
 __all__ = [
     "DataImporter",
     "DataImporterRowSession",
-    "ImportMetaBase"
+    "ImportMetaBase",
+    "ImporterExampleFile"
 ]